MSWT/pyproject.toml
[project]
name = "mswt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.16"
dependencies = [
    "aiogram>=3.21.0",
    "loguru>=0.7.3",
    "psutil>=7.0.0",
    "py-cpuinfo>=9.0.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
]
